Only chmod folders when env = development
I feel like this is inconclusive... But it depends to much on the
situation wether you want this or not. In any case: when
`$APPLICATION_ENV` is not set, the script will fail. I cannot hard-code
an environment to it, since I don't know the target env of the install.

This is the most reliable option, to only do it on development. Also, on
a production environment you might want to look into the permissions
yourself. Depends greatly on your setup.